add auth
./auth MUST be created and mpv-remote-ws won't run without it.

a sample file would be "username:password"

It's so important because mp.commandv allows using the mpv command "run".
This could be exploited from the local network or even by a remote
attacker with CSRF if tornado doesn't check the origin of the request.
(this used to be true with older versions of tornado IIRC)